removed condtions file and moved event handler into base file
